swap set_temperture and set_hvac_mode to reduce valve actuations
When the heat mode is not 'default' previously to run the thermostat controller task, the valve might have a very low local temperature for example because of venting.

When set_hvac_mode is set to auto before the temperature is fixed with the external temperature sensor the valve tends to go to 100%. After 2 Minutes this is detected by the controller, which turns it off again and restarts the valve at 0%.

After this change this should happen only in cases where there's really a large heat demand, avoiding that the fallback routine (closing the valve and starting the heating algorithm again at 0%) is used. It will still stay there to avoid that a valve can go bonkers and stay at 100% for too long.